Change id type from uint to ulong

+++ src/NTwitch.Core/Entities/IEntity.cs
@@ -3,6 +3,6 @@
     public interface IEntity
     {
         ITwitchClient Client { get; }
-        uint Id { get; }
+        ulong Id { get; }
     }
 }
